server/requirements.txt
Django>=5.2.7
djangorestframework>=3.15.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
google-generativeai>=0.8.0
psycopg2-binary>=2.9.9
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
reportlab>=4.0.0
markdown>=3.5.0
